Ignoring test for ApiProblem on HTTP Basic
This is because it doesn't work - the entry point for HTTPBasic takes over... annoyingly.
This could be fixed, but with more work than is worth for the tutorial.

Basically, the catch inside BasicAuthenticationListener should (in my opinion) just let
the AuthenticationException propagate, which would ultimatel trigger the normal
entry point (not the basic-specific entry point).